update automatic script and add git global ignore rules

+++ autoconfig.sh
@@ -1,7 +1,16 @@
 #!/usr/bin/env bash
-files=$(ls -a $1 | grep -E '^\.[^.]+' | grep -v "^\.git$")
+if [[ $# -ne 0 ]]; then
+	path=$(realpath -s "$1")
+else
+	path=$(realpath -s .)
+fi
+files=$(ls -a $path | grep -E '^\.[^.]+' | grep -v "^\.git$")
 for file in $files; do
-	rm -i ~/$file
-	ln -s $1/$file ~/$file
+	if [[ -f $file ]]; then
+		rm -i ~/$file
+	fi
+	ln -s $path/$file ~/$file
 	echo "Create soft link from $1/$file to ~/$file"
 done
+# set gitignore_global as globally ignored file rules
+git config --global core.excludesfile ~/.gitignore_global
